:recycle: spinner components 디렉토리 이동 및 분리
